+++ NEWS.md
@@ -1,3 +1,8 @@
+# eESTZwritR 0.1.0
+
++ quosures in `orderZones`
++ advice from `fieldsmakR` about removing columns. 
+
 # eESTZwritR 0.0.1
 
 + initial release